fix build

+++ tests/components/mitsubishi_cn105/climate/mitsubishi_cn105_climate_tests.cpp
@@ -5,11 +5,11 @@ namespace esphome::mitsubishi_cn105::testing {
 TEST(MitsubishiCN105ClimateTests, CelsiusTemperatureMappingMatchesExpectedValues) {
   const auto mapping = TemperatureMapping();
 
-  for (float t = 16.0f; t <= 31.0f; t += 1.0f) {
+  for (int t = 16.0; t <= 31.0; t += 1.0) {
     EXPECT_EQ(mapping.to_mitsubishi(t), t);
   }
 
-  for (float t = 16.0f; t <= 31.0f; t += 1.0f) {
+  for (int t = 16.0; t <= 31.0; t += 1.0) {
     EXPECT_EQ(mapping.from_mitsubishi(t), t);
   }
 }
